requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Авторизация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# HTTP клиент для парсинга и тестирования
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Telegram бот
python-telegram-bot==20.7

# Платежи
stripe==7.8.0

# Кеширование
redis==5.0.1

# Мониторинг и логирование
sentry-sdk[fastapi]==1.38.0

# Переменные окружения
python-dotenv==1.0.0

# Линтинг и разработка
flake8==6.1.0

# CORS (уже включен в FastAPI)
# python-cors не нужен, используем встроенный CORSMiddleware twilio
